some cleanup and refactoring
