+++ src/core/client/admin/routes/Configure/sections/Moderation/PreModerateAllCommentsConfig.css
@@ -1,3 +1,3 @@
 .specificSites {
-  padding-left: 2rem;
+  padding-left: var(--spacing-5);
 }
